Make specs pass
Capybara-accessible depends on Selenium which depends on Firefox, but Firefox
broke its connection to Capybara. So use Chrome instead. Shame on you,
Mozilla.

Capybara-webkit has a bug where FontAwesome glyphs seem to get a zero width,
so to work around it just send JS to click the New Face button.